backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Face Recognition
insightface==0.7.3
onnxruntime==1.20.1
opencv-python-headless==4.10.0.84
Pillow==11.1.0
numpy==1.26.4

# Vector Search
faiss-cpu==1.9.0.post1

# Google Drive
google-api-python-client==2.159.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0
